SQL Join
Changed SQL query to join the two tables and sort by employee ID. Changed the variables in the current try statement to include this information and altered the dictionary to store the newly added variables.

This allows us to match employee ID number, term status, termination date, etc. with the employee name.