SQL query updates and changes
Made significant changes to the SQL query that is pulling the information over.

Removed a large section of the script that was attempting to put the information in a dictionary.

Look to move forward using a pandas dataframe to store this information and work with it to compare to MS Graph.

Moved old version of code for easy access to oldversion.py