Couple UserFiltering Fixes
Some if statements to add the dictionaries that are returned during pagination from MS Graph and append the dictionaries to the aad_users list for comparison use.

Added some filtering in the ms_graph_pull() function to get service accounts and vendors out of there. There's probably a much better way to take care of that block, but it'll work for now. Hopefully can revisit that little chunk later.