95 percent to a working version
Everything should nearly be in place for a version that we can try soon in production.

Probably the most glaring thing to work on is the function to look for a user in the MS Graph information to see if they exist. There is no guarantee that the user's email will be a work email, or that their preferred name will match up or anything like that.

One possible thing we might try is sending the user's Employee ID number to AAD to match with that. In order to do that, we would probably need to add a function that would be less picky about determining if it had the right person, however.

Another outstanding item is that I need to document this really well once it's somewhat off the ground.

At any rate, it's almost done.