Possibly working.
I believe we have a toothless version completed. The functions are ready to go, they would need to be uncommented.

Final issues would be to confirm Job Titles or Departments are not going to change in a really serious way and then we should be alright to try a run at this thing.